Prevent YUI_COMPRESSOR_CHECK if no JS files present

+++ travis.script.sh
@@ -18,8 +18,8 @@ if [ -e phpunit.xml ] || [ -e phpunit.xml.dist ]; then
 fi
 
 # Run YUI Compressor Check
-if [ "$YUI_COMPRESSOR_CHECK" == 1 ]; then
+if [ "$YUI_COMPRESSOR_CHECK" == 1 ] && 0 != $( find $PATH_INCLUDES -name '*.js' | wc -l ); then
 	YUI_COMPRESSOR_PATH=/tmp/yuicompressor-2.4.8.jar
 	wget -O "$YUI_COMPRESSOR_PATH" https://github.com/yui/yuicompressor/releases/download/v2.4.8/yuicompressor-2.4.8.jar
-	java -jar "$YUI_COMPRESSOR_PATH" -o /dev/null $(find $PATH_INCLUDES -name '*.js') 2>&1
+	java -jar "$YUI_COMPRESSOR_PATH" -o /dev/null $( find $PATH_INCLUDES -name '*.js' ) 2>&1
 fi
